Create bucket for owner directly instead of linking it.
Use the credentials of the specified user to create a bucket instead of link that user to a bucket that was created via the admin account.

This approach will workaround the issue https://github.com/aquarist-labs/s3gw/issues/86 to be able to create buckets, but updating a bucket, e.g. linking it to another user, is still broken.

Signed-off-by: Volker Theile <vtheile@suse.com>