Support sticky header in datatable pages
Stick the table column header to the top of the nearest scrolling ancestor and containing block. Defaults to `true`.

In the beginning it was planned to treat the action-buttons accordingly, but this was discarded because it can be assumed that the actions provided by the action-buttons (above the table) are not in the focus of the user when scrolling the table rows. Individual actions can still be performed using the actions-menu in the last column of each row (if added by the developer).

Signed-off-by: Volker Theile <vtheile@suse.com>