backend: drop some type stubs, add custom typings
We were installing 'types-aiobotocore-lite' because the 'session.pyi'
provided by 'types-aiobotocore' was not enough, for some reason.
However, during 'pip install' we do not control the sequence libraries
are installed, and again and again we would have the latter being
installed after the former, effectively squashing the good
'session.pyi'.

Instead, we grabbed the 'session.pyi' we need and placed it in the
'typings/aiobotocore/' directory, dropping the 'types-aiobotocore-lite'
dependency altogether. We carry more stuff, but at least we can ensure
some consistency.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>