Inconsistent restore behavior
The datatable `Restore` action button does not behave the same way as the row action menu. In this case the `Restore` button is not disabled when the selected item is a delete marker or is the latest version.

Additionally the `Download` table action button will be disabled when the selected version is a delete marker.

Fixes: https://github.com/aquarist-labs/s3gw/issues/747

Signed-off-by: Volker Theile <vtheile@suse.com>